pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.6",
    "docling>=2.64.1",
    "easyocr>=1.7.2",
    "flask>=3.1.2",
    "groq>=0.37.1",
    "paddleocr>=3.3.2",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.6",
    "pypdf2>=3.0.1",
    "rank-bm25>=0.2.2",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
